fix failing itest

+++ src/main/java/com/skthvl/cinemetrics/model/dto/AddRatingDto.java
@@ -1,6 +1,7 @@
 package com.skthvl.cinemetrics.model.dto;
 
+import java.math.BigInteger;
 import lombok.Builder;
 
 @Builder
-public record AddRatingDto(int rating, Long movieId, String userName, String comment) {}
+public record AddRatingDto(int rating, BigInteger movieId, String userName, String comment) {}
